backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
openai==0.28.0
python-dotenv==1.0.0
email-validator==2.0.0.post2
psycopg2-binary==2.9.7
alembic==1.12.0 